more work on printf

+++ kclib/_stdlib_numconv.c
@@ -6,6 +6,7 @@
 #include <errno.h>
 #include <ctype.h>
 #include "intinc/shmath.h"
+#include "intinc/stdlib.h"
 
 int atoi(const char *nptr){
 	return (int)strtol(nptr, (char **)NULL, 10);
@@ -111,7 +112,7 @@ size_t __countnumeric(const char** ptraddr, int* base, bool* minus, const char**
 	return validlen;
 }
 
-uintmax_t __radixval(char c){
+uint8_t __radixval(char c){
 	static char val_radix_upper[] = {
 					'0', '1', '2', '3', '4', '5', '6', '7', '8', '9',
 					'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X', 'Y', 'Z'
